Theoretically, fix random seed
But, practically and unfortunately, don't...

Theory was that, since OUTPUT_PIN makes its way
to GND/LOW eventually, it isn't floating enough
to provide different HIGH/LOW values. And that's
why, when using batteries but surprisingly not
when using the programmer for power, we always
seemed to get the same first interval sequence.

UP_PIN and DOWN_PIN seem better options, since
they float before the internal pull-downs are set
and assuming user isn't holding their buttons.

However, changing to UP_PIN yielded identical
behavior.

Not only that, but straight up using LOW did too.

So, IDK wussup, but theoretically-improved-but-
practically-identical is better enough in my book.
Committing with this lengthy message as proof of
effort.